requirements.txt
colorama==0.4.6
types-colorama==0.4
python-dotenv==1.0.1
pymongo==4.8.0
termtables==0.2.4
pycryptodome==3.20.0
pycryptodomex==3.11.0
